userportal: not passing hostsubject to spice-xpi
The problem was, that the add_object_column_white_list_table.sql
was ran as pre-upgrade script, but it tried to fill the
object_column_white_list according to the not-yet crated
view named 'vds'.

Solved by creating also post-upgrade scripts, as a last
step of the upgrade process and making the
add_object_column_white_list_table.sql to be a
post-upgrade script.

Change-Id: Ice7c00a158a8778a6cb424be4e9720715a4df283
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>