build: place proper name for jars/poms if package name modified
Per Juan notes at[1]

According to the Fedora packaging guidelines the names of the POMs
should start with JPP.%{name}. See [2].

However, there is dependency between the jar location and the prefix
of the pom file.

The solution is to create symbolic link at _javadir if package name
is different from upstream name.

[1] http://gerrit.ovirt.org/#/c/6252/5/packaging/fedora/spec/ovirt-engine.spec.in,unified
[2] https://fedoraproject.org/wiki/Packaging:Java#maven3.

Change-Id: Iae5434a00537892a21446f18c28b263683402d09
Signed-off-by: Alon Bar-Lev <alonbl@redhat.com>