core: Extend the lapse of timeout
The unit-test fails selectively when trying to validate the expiration
time of a specific value.

The value is set with time of the system and compared to another call
for that system time.

The suspicious is that the second call might return a lower time than
the first one, as there is no contract for accuracy.

The solution is to increase the lapse of timeout.

Change-Id: I330a9b1ff5c20b0885f262943d1fe2107f8da7c7
Signed-off-by: Moti Asayag <masayag@redhat.com>