packaging: Ports 8006, 8007 and 8008 are not required
The message presented by engine-setup when the installation
finishes states that the ports 8006, 8007 and 8008 have to be open
in the firewall, but we don't use them for anything currently. The
only ports we need to open in the firewall are the HTTP and HTTPS
ports, as selected by the user.

Change-Id: I77be99f09e8b392c337739d5b88d028dbdc73e77
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>