restapi: removing access protocols and acl in create volume
AccessProtols and AccessControlList(ACL) will inturn converted
to volume options nfs.disable and auth.allow respectively.

As passing option also supported while creating a volume,
it creates confusion when some use both acl and auth.allow.
If different values are provided for them, it is difficult to
decide which one gets the precedence.

Both AccessProtols and AccessControlList is removed from the
Volume resource to avoid the conflicts.

Change-Id: Ib0dad83cc2e3dc0dfb698677084150d870289eff
Signed-off-by: Kanagaraj M <kmayilsa@redhat.com>