userportal,webadmin: Blank template has active action buttons (#840856)
https://bugzilla.redhat.com/840856

Have at least one non-blank template. Select first the non-blank,
than the blank. The action buttons will be enabled.

Fixed by enriching the login of
TemplateListModel.isBlankTemplateSelected() which now checks
all the selected templates (getSelectedItems()) and not only
the first selected (getSelectedItem()).

This logic is common for both webadmin and userportal,
so this fix fixes the same problem in userportal as well.

Change-Id: I3959752d734c01d6d1b20afeacbecaa59f2b6608
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>