webadmin: Export shows incorrect warning(#840854)
https://bugzilla.redhat.com/840854

Have 2 VMs/Templates which are already exported to an
export domain. Select them again and press export.
In the export dialog a warning containing only one of
this VMs/Templates is shown.

Fixed by rewriting the algorithm composing the message
in VmBaseListModel. This logic is common for both
VM export and Template export, so this fix fixes
this issue for both exporting VMs and Templates.

Change-Id: I84d4103fd2d2b4292575cd8b51839ee18cf42352
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>