webadmin: New pool contained template provisioning (#839046)
https://bugzilla.redhat.com/839046

If there are no templates in the environment, than:
Pools->New->Resource Allocation
It contains the Template Provisioning.
The template provisioning makes actually no sense for the
pool dialog, so it never should be there.

It was there, because the logic which disabled it was in the
Template_SelectedItemChanged method which is called when the
selected template changes. The problem is, that when there is
no template in the system at all, this method gets never called.

Fixed by moving this logic into the Initialize method.

Change-Id: I5ade45b807f35161511e352ebd0ffb818730fcee
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>