engine: Edit Pool shows different RAM than defined (#833829)
https://bugzilla.redhat.com/833829

Have a template X with 256 MB memory, than:
Pools->New->select template X->change RAM to 512 MB->save
Than select the newly created pool and Edit.

The memory will be 256 MB.

The problem was, that both the
(New/Existing)PoolModelBehavior classes let the filling of the
window model to the parent class (PoolModelBehaviorBase) which
used the template. It is correct for the
NewPoolModelBehavior but not for the ExistingPoolModelBehavior.

Fixed by generalizing the logic of setting up the window model
to take VmBase instead of VmTemplate and calling this from the
specific classes (New/Existing)PoolModelBehavior with the
correct parameter.

Change-Id: I7fbc3774cdf4ed4209724fd8d5053bbef0a76f7b
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>