webadmin: edit pool dialog freezes when when SD is down (#839594)
https://bugzilla.redhat.com/839594

Have a pool on a storage domain which is down.
Go to pool tab, press edit. The window will
stack forever.

Fixed in a way, that the dialog opens the same way as if
the SD would not be down. When the OK button is pressed,
an error comes back from the server.

Fixed by avoiding NPEs and ClassCastExceptions on different
places.

This patch also refactors the Edit method to use the
GetVmDataByPoolId query, which is more efficient than the
AsyncDataProvider.GetVmList and selecting the first element.

Change-Id: Ic9635873a04c71bdf94a011a27dd47cb5c9b259b
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>