engine: Add sync networks to Setup Networks (#838300)
https://bugzilla.redhat.com/838300

Added list of networks that need to be synced to SetupNetworks
parameters. The list allows to specify which networks should be
synched.

Setup Networks will now sync the network if it needs to be synched (i.e.
specified on the list), but if network is unsynched then it's not
allowed to be edited until it's synched.

Change-Id: I433e684f09b104c8eda82b21129bba3e019c58d6
Signed-off-by: Mike Kolesnik <mkolesni@redhat.com>