core:pre upgrade scripts will not run ...(#842119)
pre upgrade scripts will not run correctly upon 3.0 DB restore

https://bugzilla.redhat.com/show_bug.cgi?id=842119

Restore a 3.0 database and running the 3.1 upgrade scripts on it is not
working.

This patch fixes this problem by :
1) force pre-upgrade scripts to run if any upgrade file is about to run
, this will be done even if no change was detected in views, sp &
upgrade files as it may occur when the database is restored.

2) make pre-upgrade and post upgrade scripts ordered since the
configuration settings and some fixes in the custom sql file must run
first. This is done by prefixing the script name with a 4 digit number.
(0000 00010 00020 etc.)

Change-Id: Ie23fba63cee81c5a5b3f1ad4076d627a06614fc0
Signed-off-by: Eli Mesika <emesika@redhat.com>