core: Animal-sniffer to check JDK6 compatibility
Since it was agreed that oVirt should remain backwards compatible to
JDK6 APIs, it is needed to introduce a mechanism to validate this.

Animal Sniffer is a standard maven plugin that enforces code
compatibility to jar versions (including but not limited to JDKs),
without requiring these jars to be installed - only a signature spec is
required, and Animal Sniffer provides such signatures to older JDK
versions.

Change-Id: I83ffd34058ae3862087725acb1bcc21b1b4dcd38
Signed-off-by: Allon Mureinik <amureini@redhat.com>