core: Host does not move to non-operational (#822451)
https://bugzilla.redhat.com/822451

If an interface of required network (by cluster) is reported from VDSM
as 'down', the host should move to non-operational.

The patch modifies a problem of overriding broken status with a correct
one, and loop for all of the interfaces to collect more information
about additional problematic networks and interfaces.

Change-Id: I8911ed41e2f9e01451c58b70b0679c6c606ea759
Signed-off-by: Moti Asayag <masayag@redhat.com>