core: QueriesCommandTest: removed wrong test
Removed a test in QueriesCommandBase that essentially checked that
the construction of a query from VdcQueryType works.

First, this test should not be here.

Second, since several queries init static blocks that depend on external
resources (that should be mocked, but aren't here), it may cause other
tests to fail.

Change-Id: I9f8d5df49dbe78a8288021567907e31ff53a75a3
Signed-off-by: Allon Mureinik <amureini@redhat.com>