core: Preserve 404 error code in error page (#842653)
https://bugzilla.redhat.com/842653

We used to redirect to the error page in order to avoid problems
with relative URLs, but this is causing problems with clients (GWT
for example) that follow the redirect and this don't recieve the
error code. This patch replaces the redirect with forward and
fixes the error page (the index page in fact) so that it doesn't
use relative URLs.

Change-Id: I84001b195c06ae8be38e500ba687353080d04830
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>