core: fixing import storage domain to save additional NFS options (#835961)
https://bugzilla.redhat.com/835961

This patch uses the AddStorageServerConnectionCommand in the import NFS
storage domain (same as when adding a new domain), instead of using
ConnectStorageServerVDSCommand, as it saves the server connection in the
database. Also, it puts the additional NFS options in the server
connection object as well, so that they will be persistent.

Change-Id: I87ac4a2b8a7f2083f2d5310cc9899514936d6111
Signed-off-by: Oved Ourfali <oourfali@redhat.com>