engine: Edit Pool shows different RAM than defined (#833829)
https://bugzilla.redhat.com/833829

Have a template X with 256 MB memory, than:
Pools->New->select template X->change RAM to 512 MB->save
Than select the newly created pool and Edit.

The memory will be 256 MB.

The problem was, that both the
(New/Existing)PoolModelBehavior classes let the filling of the
window model to the parent class (PoolModelBehaviorBase) which
used the template. It is correct for the
NewPoolModelBehavior but not for the ExistingPoolModelBehavior.

Fixed by generalizing the logic of setting up the window model
to take VmBase instead of VmTemplate and calling this from the
specific classes (New/Existing)PoolModelBehavior with the
correct parameter.

To enable this generalization, getnum_of_cpus() method has been
moved from VmTemplate to VmBase.

As the setter was not needed anymore because the CXF is not
used anymore, the setter has been removed from the VmTemplate.
This removal enabled also to remove the "num_of_cpus" from
ObjectIdentityChecker in the VmTemplateHandler.

Change-Id: Idaa1a92f7fe85f9c604da8916be5672d8d5b1e8b
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>