restapi: Fix matching of empty list
If empty list is expected in test, then the matcher would fail because
it contains no elements and the default return value is initialized to
'false'.
Changed default initialization of the return value to 'true' so that it
will be changed to false in the objects are found inequal. No other
change is necessary in this regard, since the method already handles it.

Change-Id: If9144c448881d9373c22c69bd7d222ab88b59f2a
Signed-off-by: Mike Kolesnik <mkolesni@redhat.com>