packaging: engin-setup - Drop DB before restore when upgrade fails (BZ#844341)
When the DB upgrade fails, the setup tries to restore the backed up DB.
In a specific case, when the upgrade succeeded, but engine-config
execution fails, the DB restore fails, since the DB already exists.

This patch fixes the restore flow, by adding a call to db drop method.
Also, it fixes the reason for the engine-config setup, and always
configure the db connection string and password.

Change-Id: Ic465815c683c0e4241f520960d4acc0dfa59f95a
Signed-off-by: Ofer Schreiber <oschreib@redhat.com>