core: force update VM to down
In case unlock VM is called, the action should be performed with
no consideration to the vm that passes as a parameter.
this vm may be cached. so we want to force update.
if the command is called and in th vm in the DB is already down, that is
another problem (why we called the method).

Change-Id: I01cb0b7c939ee506d752b34dc0d7a16e46e36687
Signed-off-by: Gilad Chaplik <gchaplik@redhat.com>