webadmin: modify UI in GuideMe dialog (#839316)
https://bugzilla.redhat.com/839316

When a user adds a new Data Center and follows the "Guide me" dialog,
after adding a cluster, in the next dialog the "select host" button
should appear but it doesen't.

The UI was modified to match older UI version - so the button would always
appear. Neverthelass, if the host version is not compatible with the
cluster version (or no host in maintanance mode) the button will be
disabled.

Change-Id: If898016a594c1f4d40440d7f5e290430fe0e2ea4
Signed-off-by: Ofri Masad <omasad@redhat.com>