core: Can remove proxy for host but left entry at DB (#844306)
https://bugzilla.redhat.com/844306

The problem described at bug is very rare and very difficult for reproduce
but , it can cause for serous trouble at engine side
The problem is following:
1. We are removing host
2. The removing of host and removing its proxy done inside same transaction
3. Transaction was rollbacked due to some error
4. Entry at dabase left, proxy was removed
Result: we have an orphan entry of host at DB which we can not remove, activate , etc.
Our process can try to choose that entry as SPM or try to activate via Auto recovery, it is obvious
that we will not success.
Solution: Diasable global transaction, perform removefrom DB  at separete transaction if it fails, the proxy will
left, if it success the proxy will be removed.

Change-Id: I79610be56eab28a74627db66846d41702b9a440d
Signed-off-by: Michael Kublin <mkublin@redhat.com>