webadmin: fix unable to save priority (#840006)
https://bugzilla.redhat.com/840006

When a user try to create new vm or edit an exsisting one, the priority
is not saved.

The max priority value and the VM priority value where bothe cached to
the same key. once the vm priority was 1 - the max priority was 1 and
the 3 integer options for choosing priority where all equal 1.

Change-Id: I863fa7a00c3208d6137bd982f9702fe850d42bed
Signed-off-by: Ofri Masad <omasad@redhat.com>