restapi: Update Storage-Domain - handle type
restapi does not have 'Master' data type. A storage-domain of type master is
simply displayed as data. When user updates a storage-domain, he might send
back <type>data</type>. Backend would fail this, because it looks like the
user is updating the type field, which is forbidden (because the old value
is 'master', and user gave 'data').

This patch intercepts the update and replaces 'master' back to 'data' if needed,
so that the backend won't fail on ilegal update

Change-Id: Iabaaa2d262a55e073d4aebe32ac76bbc9ab92d0d
Signed-off-by: Ori Liel <oliel@redhat.com>