userportal: removing a user from 1 pool - UP fails
When a user has rights on a VM which is a part of a pool,
which he does not have rights to, than no pools will be shown
on the UP at all.

The problem was in VmItemBehavior.UpdateProperties() on line 172.
The problem is, that when there is a VM which has a pool, but the
user does not have rights on this pool, the frontend throws an
exception. The pool was used only to set up the VmPoolType.

Fixed by removing this logic, as the VmPoolType does not make sense
for the VM (and is not used anywhere), so does not need to be set up.

Change-Id: I030f89447df4a75dca3e52a3986e2a07a086a677
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>