webadmin: Main events tab takes too long to fetch data
The same issue as already fixed for subtabs in:
eda0ef6be8ba97e6a835907a998683cdca1fcb74
The problem for main tab is that the main tab does not have the
entity set, so the refresh command was not executed.

Fixed by calling the refresh without the check on the
entity in the EventListModel and creating a new class
SubTabEventListModel which is a common parent for all
sub tab events and adding this check only here, so it will
apply only for the sub tabs.

Change-Id: I7d0d0df896113c2989a1864f2ca89d488a63e446
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>