core:Remove multiple columns in group by to use view
The following patch removes group by which is used for disks.
The group by is being done to use array for storage for disk, and it is
done on three columns.
Since the view use many columns each time a new column is added we need
to add it to the group by, which cause the views to be large and ugly.

The suggested fix is to use a view specifically for storage which use
those three columns, and use join on it when ever it is used.

Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>
Change-Id: I273c423009f4718bc527a6155745fda757771303