core: fixed NPE during removal of floating disk (#845046)
https://bugzilla.redhat.com/show_bug.cgi?id=845046

when trying to remove a floating disk NPE was encountered
beacuse of the use of NGuid.createGuidFromString (who can return null)
instead of Guid.createGuidFromString who on DiskImageMapper. The current
change provides null check and equality check to Guid.Empty in order to
not affect other flows that count on null on that property.

Change-Id: I6ffca77cc5a02558847366dd8ccb886ee515d612
Signed-off-by: Liron Aravot <laravot@redhat.com>