core: misleading error when trying to fen(#843168)
ovirt-engine-backend: misleading error when trying to fence host without
PM configured

https://bugzilla.redhat.com/843168

Problem was that the code in the canDoAction of the base class was not
adding any message if fencing is not configured in the host and also
retries the operation event though there is no chance that it will
succeed.

Changing the code to add a message when Start/Stop is called on a non
configured Host.
Make StopVdsCommand call the super.canDoAction in the same manner as
StartVdsCommand does.

Change-Id: I7aab4f6a7c70fa5e18db106ea15a7b20b709db28
Signed-off-by: Eli Mesika <emesika@redhat.com>