webadmin: Bad formatting in dialog (#835404)
https://bugzilla.redhat.com/835404

When the user trys to "Configure local storage" and go to memory
optimization sub tab he gets a corupted message contining missing
formating values.

i.e.: "Memory Page Sharing Threshold set to %1$s"
instead of: "Memory Page Sharing Threshold set to 150%"

The formating was not handelled when the item was not changed. So the
message was not formated correctly.

Change-Id: I36c34ce25b50697de7c1ee0165f13c05616605be
Signed-off-by: Ofri Masad <omasad@redhat.com>