engine: Change violations to Map (#844158)
https://bugzilla.redhat.com/844158

Changed violations list to be a Map in order to allow generating a list
of offending entities (which will be generated by the pattern
"{KEY}_LIST" where {KEY} is the violation name).

Also added a method to add a violating entity to the map for following
patches to start using.

Change-Id: Ic70f4d93e1ca06a9991e6e443e92f6be11fc88d4
Signed-off-by: Mike Kolesnik <mkolesni@redhat.com>