webadmin: new VM: pin to host not used from template(#845570)
https://bugzilla.redhat.com/845570

New VM -> select template -> host tab

The data are not read from the template.

The probelem was that the reaction to the template selection
change was not the read of the update of the host pinning props.

Fixed by extraction of the respective logic and added to the
VmModelBehaviorBase (methods doChangeDefautlHost() and
updateHostPinning()). This methods are called also from the
Existing* behaviours to read the respective data from the
VM (or VM Data in case of Pool).

This patch also enriches the AbstractVmPopupWidget by
adding the listener to IsAutoAssign property, because
this is not connected with the model itself.

Change-Id: I561c5ee3967676ffd89758820b09a192a741a2e8
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>