core: add ability to run shell upgrade steps
Up to now, we were able to run only SQL files in upgrade pre_upgrade and
post_upgrade.
This patch adds the ability to add shell scripts as upgrade steps or as
part of the pre/post upgrade.
Shell script files preserve the same naming conventions as regular SQL
upgrade scripts (extension is .sh) and are recorded in the same way to the
schema_version table.

This patch is mandatory to some bugs that requires upgrading the
database via a shell script.

Change-Id: Ia3a88c21c97ebdd9bd35c2b0da87cba1cbde54c7
Signed-off-by: Eli Mesika <emesika@redhat.com>