tools: Add tolerance to engine-setup min memory threshold
engine-setup exits with error if system is under memory
threshold in basedefs.py (CONST_MIN_MEMORY_GB).
Since engine-setup is using 'free -g' command, using
default threshold (2GB), systems with less than 2048 MB
of RAM will be classified as 1 GB.
e.g. a system with 2048 MB with some memory allowed
to video cards will not be able to run engine setup.
This patch will change engine-setup to use 'free -m',
change the memory threshold to be in MB and add a
tolerance of 5% to MIN_MEMORY.

http://bugzilla.redhat.com/843621

Change-Id: Ibd22cccc85c62535c57b651347bd6e7ccedec625
Signed-off-by: Amador Pahim <apahim@redhat.com>