core: use allow viewing children only in disks, templates and VMs (#846300)
https://bugzilla.redhat.com/show_bug.cgi?id=846300

This patch is one of the patches for fixing the bug above.
Previous commit added the ability to specify whether an action group
allows to view the children on the granted object or not.

This ability to specify that is relevant only for specific object types, which are
disks, templates and VMs, so this patch removes it from all the other
objects in the system.

Change-Id: I4dfecc6e433f13e773532864fafd44d9ffe8c7d0
Signed-off-by: Oved Ourfali <oourfali@redhat.com>