core: Make NetworkOperationsConnectivity configurable (#840746)
https://bugzilla.redhat.com/840746

The patch exposes NetworkConnectivityCheckTimeoutInSeconds configuration
value via the engine-config tool and assign to it valid values of:
10-160 seconds.

The lower boundary (10) designed to prevent preliminary rollback.
The upper boundary (160) designed to remain a reasonable period for
rollback before the default value of vdsTimeout (180 seconds).

The last assumption is made since there is no support dependency between two
configuration values. If a user increases the default value of vdsTimeout,
 same can be done for NetworkConnectivityCheckTimeoutInSeconds by extending
the valid value as specified in /etc/ovirt-engine/engine-config/engine-config.properties

Change-Id: Ie8425e3a10636f5a409c6748f6f17858c3ff666c
Signed-off-by: Moti Asayag <masayag@redhat.com>