core:encrypt vds_static pm_password field on upgrade to 3.1
In 3.0 all host having PM configured add the agent password
(pm_password) in plain-text.
In 3.1 when we configure PM, we also encrypt its password and it stored
in the database.
This patch goes over all hosts that are upgrading from 3.0 to 3.1 and
encrypt the agent password in the database (vds_static.pm_password) if
it has PM configured (pm_enabled = true)
The patch gets all information needed for the encryption from the
database configuration values (stored in vdc_options) table) and uses
the encryption exposed by store-utils to get the work done.

All the above is executed via a shell script
The script is numbered as 03_00_0420 since 03_00_0410 was the last 3.0
script and we want the shell script to run only when upgrading from 3.0
to 3.1.

Also, please note that the pm_password field in vds_static was changed
from varchar to text in 03_01_0030 and we had to verify that this change
occurs also in our patch. This is OK since the general function used for
that to change the type validates the source type.

Change-Id: I4196ef1070ffabdbc7b8b2846fc4f0340ef07ac8
Signed-off-by: Eli Mesika <emesika@redhat.com>