core:Print clusters when upgrading DC version(#844425)
https://bugzilla.redhat.com/show_bug.cgi?id=844425

When we upgrade Data-Center we check if all cluster also has a validate
level (which is the same as the DC or higher), if not, the user gets a
message that he should upgrade the clusters, but there are no mention
which clusters it should be upgraded.

The proposed fix, adds the cluster names that should be upgraded to the
message, and add a test to cover it.

Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>
Change-Id: Ia9c92d1c477354f7f5fa66895ecf2ab38c10bdf6