core: Add Network Filtering rules to VMs
The patch introduces the Network Filtering feature.
The feature is enabled for 3.2 compatible cluster level.
If the feature is enabled, the engine dictates libvirt via
VDSM to add a network filter rule to prevent VM mac spoofing.
A single filter is supported per vnic. The filter is set on vnic
during starting a VM or when hotplug of vnic is activated.

Feature page:
http://wiki.ovirt.org/wiki/Features/Design/Network/NetworkFiltering

Change-Id: I5776ea730ef1733fe4f9efe852c07053573dc5ba
Signed-off-by: Moti Asayag <masayag@redhat.com>