core: Redirect to index when page not found
This is a fix for commit 2107a03979 (ouput index.html as error page),
which was a fix for commit 66edaab361 (redirect servlet).

Introduce error.html at root, webapp, userportal-gwtp. Use HTTP meta
refresh to redirect to root context.

As a result the index.html hrefs can be relative as they should.

Repeating the note:

HTTP Redirect causing problems with clients (GWT for example)
that follow the redirect and this don't recieve the error code.

Change-Id: I63e7992a748935347eee7b65930bc6d29fdb92db
Signed-off-by: Alon Bar-Lev <alonbl@redhat.com>