webadmin: merge the 2 SPM columns into 1 (#845320)
https://bugzilla.redhat.com/845320

The two columns - "SPM Status", "SPM Priority" were merged.
New column is called "SPM"
For a host which is Contending or SPM - the status is shown.
For other host - the priority is shown (Low, Normal, High, Custom)

The numberes behind the priority were drop from the column and the 'Edit
Host' -> SPM subtab.

Change-Id: I1db81336eac01ea2e713912a37e5bec90e10dfd5
Signed-off-by: Ofri Masad <omasad@redhat.com>