engine: We are not blocking reconstruct on domains during spm recontend (#843100)
https://bugzilla.redhat.com/843100

The problem was caused by race which was created by simentenious run of maintaince master storage domain
and spm host.
In order to solve a problem a locks will be used.
A shared lock on pool will be use at MaintananceNumberOfVdssCommand if we are trying to maintaince spm.
An exclusive lock will be used in DeactivateStorageDomainCommand if we are trying to deactivate master domain
and shared lock on any other data domains.
The patch contains a fix of possible NullPointerException at DeactivateStorageDomainCommand
Change of order moving to maintaince of hosts at MaintananceNumberOfVdssCommand, in order to move spm a last
to maintaince.

Change-Id: Ia929d8cfab5904cb8e66d18cf50e037d36d8d373
Signed-off-by: Michael Kublin <mkublin@redhat.com>