webadmin,userportal: filter quota in quota dropdown (#847511)
https://bugzilla.redhat.com/847511

quota drop down should be populated with the backend query,
but in case the list of quota is missing the current quota of the object,
it will be inserted to the list automatically.
mainly used in edit and in the user portal:
edit an object (e.g. vm) the object has a quota that the user doesn't have
consume quota on. so the list that will be fetched from the server won't have that quota,
so in the ui we will insert it to the list and select it by default

Change-Id: I823c82c00f632258f7b885ae2ef5a7cb833a0762
Signed-off-by: Gilad Chaplik <gchaplik@redhat.com>