core: Don't erroneously remove template->storage mappings
If a storage domain is force removed and a template has a disk on that
SD, the image is removed regardless of whether that image also resides
on other SDs.  In the case where the template disk has been copied, the
template and images should remain intact and only the references for the
specific SD being removed should be deleted.

Change-Id: I058d160622990d2b42ccc02a61551c6f61550ade
Signed-off-by: Greg Padgett <gpadgett@redhat.com>