engine: Exception when updating lun disk properties (#844153)
https://bugzilla.redhat.com/844153

When updating properties of a LUN disk, the audit logger attempted to
retrieve the storage pool id of the disk.  The method to return the id
cast the Parameter to a DiskImage, even though the Parameter was a
LunDisk, causing a ClassCastException.

Because the storage pool isn't in the audit log, remove the method
UpdateVmDiskCommand.getStoragePoolId() altogether.

Change-Id: Iba86289c25937002c9c07a06db5651bca4485d38
Signed-off-by: Greg Padgett <gpadgett@redhat.com>