webadmin: Import gluster cluster
Cluster dialog is extended to support import an
existing gluster cluster. The user has to enter the
details of any server(host) in the gluster cluster to import.

The flow is

1. At the time of creating a new cluster the user enables
Gluster service and selects import existing cluster configuration.

2. Provides the details (hostname/IP, password) of any server
in the existing gluster cluster

3. When the user lose focus on the hostname/IP field, a rpc call is
fired to fetch the fingerprint of the host

4. User validates the fingerprint and clicks on Ok in cluster dialog.

5. A new dialog will be shown with all hosts in the gluster cluster

6. User validates the fingerprint of all the hosts and provides passwords.
Clicks on ok, cluster will be created and the hosts will be added
to the cluster.

Change-Id: I502acdb461c0b8557be2f933c7fd6ba5da469f23
Signed-off-by: Kanagaraj M <kmayilsa@redhat.com>