core: live snapshot failes in case of a deactivated disk (#845020)
https://bugzilla.redhat.com/845020

when creating live snapshot for a VM, a SnapshotVdsCommand was executed
for all disks, plugged and unplugged in order to switch the images that
the vm is running on to the newly created images.  in case of an
unplugged disks,
the command had failed as the VM doesn't currently use the unplugged
disk. this fix
filters out the unplugged disks from the list of disks sent to the
SnapshotVdsCommand.

Change-Id: I27c32525794b4ec3c9d8c78f033a3c9d39da72b9
Signed-off-by: Liron Aravot <laravot@redhat.com>