core: Fix calculations of snapshot quota (#848028)
https://bugzilla.redhat.com/848028

Fix 1: The sized used for quota calculation of each snapshot was fixed.
The former value was the size of the predefined disk, the new value is
the actual size the snapshot took in storage.

Fix 2: The snapshot-preview did not pass the quota ID to the new active
image. After the commit, the active image was linked to a null qouta and
so a new snapshot could not be taken.

Change-Id: I6cf5d23a9c5c90b64b330cadf4f7c1b67fad45fa
Signed-off-by: Ofri Masad <omasad@redhat.com>