packaging: Retry failed AJP connection after 5 seconds
The configuration that we currently use to connect to the
application server using AJP is the folliwing:

  ProxyPass / ajp://localhost:8009/

This means that the web server will use the default value of 60
seconds for the "retry" parameter. This parameter controls how
long will the web server wait before retrying a failed AJP
connection. For us it means that if we do something like "service
ovirt-engine restart" the web server will force us to wait 60
seconds regardless of how long it takes to restart the engine. In
environments with several application servers this makes sense,
but with only one application server it is better to wait less, or
nothing. This is what we do in this change:

  ProxyPass / ajp://localhost:8009/ retry=5

This means that the web server will wait 5 seconds before
establishing a new connection after a failure.

Change-Id: Ia8291e0671c2feab0e0ae2878795ad5be2c17cc4
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>