core: Introduce ChangeQuotaCommand - disks (#848310)
https://bugzilla.redhat.com/848310

We want the abillity to change quota for disks and vms.
added an abstract command (ChangeQuotaCommand), that's resposnible for that.
Currently implemented the ChangeQuotaForDisks, which gets the disk and quota
and tries to change the disk quota, ChangeQuotaForVms will be added in the near
future.
Added a test to ImageDao, that checks the image quota actually changes.

Change-Id: I0017a18109faecd1acab52897bba5474b6600390
Signed-off-by: Gilad Chaplik <gchaplik@redhat.com>