engine: Obtain old network name from correct NIC (#849971)
https://bugzilla.redhat.com/849971

The old network name should be taken from the NIC it's defined on.
If there is more than one network defined on the given NICs, this
command can't handle it and should fail.

VLANs are also handled since a NIC can have several VLANs based on it
which in turn can have networks attached to them.

Change-Id: I1311be342aa72bac9f1b79cec816be52f37d966d
Signed-off-by: Mike Kolesnik <mkolesni@redhat.com>